requirements.txt
click==8.1.7
PyYAML==6.0.2
pydantic==2.9.2
pydantic-core==2.23.4
jsonschema==4.23.0
markdown==3.6
pypdf==5.0.1
watchdog==4.0.2
cryptography==43.0.1
pytest==8.3.2
ruff==0.1.9
mypy==1.7.1

pillow==10.1.0
